pyproject.toml
[tool.poetry]
name = "tfg"
version = "0.1.0"
description = ""
authors = ["Carlota de la Vega <carlotadlavega@gmail.com>"]
readme = "README.md"
packages = [
    { include = "tfg" }
]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.0"
torchvision = "0.17.0"
torchaudio = "2.2.0"
requests = "^2.32.3"
Pillow = "^11.1.0"
numpy = "<2.0"
pytorch-pretrained-biggan = "^0.1.1"
scipy = "^1.15.0"
nltk = "^3.9.1"
diffusers = {git = "https://github.com/huggingface/diffusers.git"}
transformers = "^4.45.2"
accelerate = "^1.0.1"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
ruff = "^0.9.5"
datasets = "^3.2.0"
peft = "^0.15.0"
bitsandbytes = "^0.45.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
